Uploaded API to query CDMS
This API will accept the database name and molecule tag as inputs and return the frequency, uncertainty, and intensity of all spectral lines emitted by the molecule with the specified molecule tag.  Currently, it only works for the CDMS database.  We still need to implement JPL, OSU, SLAIM, and TopModel (at the minimum, possibly more).